+++ src/structlog/_base.py
@@ -153,7 +153,7 @@ class BoundLoggerBase:
         for proc in self._processors:
             event_dict = proc(self._logger, method_name, event_dict)
 
-        if isinstance(event_dict, str):
+        if isinstance(event_dict, (str, bytes)):
             return (event_dict,), {}
         elif isinstance(event_dict, tuple):
             # In this case we assume that the last processor returned a tuple
@@ -163,9 +163,9 @@ class BoundLoggerBase:
             return (), event_dict
         else:
             raise ValueError(
-                "Last processor didn't return an appropriate value.  Allowed "
-                "return values are a dict, a tuple of (args, kwargs), or a "
-                "string."
+                "Last processor didn't return an appropriate value.  Valid "
+                "return values are a dict, a tuple of (args, kwargs), bytes, "
+                "or a str."
             )
 
     def _proxy_to_logger(